requirements.txt
pandas
numpy
geopandas
shapely
streamlit
plotly
scikit-learn
pyyaml
openpyxl
requests

joblib
